requirements.txt
requests==2.31.0
setuptools==69.0.2
PyYAML~=6.0.1
pip~=23.3.1
six~=1.16.0
python-dateutil~=2.8.2
chardet~=5.2.0